# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "409":
          description: Conflict - duplicate phone/email/whatsapp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - duplicate phone/email/whatsapp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suppliers/{id}/payment-methods:
    get:
      operationId: getSupplierPaymentMethods
      tags: [supplier-payment-methods]
      summary: List payment methods for a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierPaymentMethod"
        "404":
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSupplierPaymentMethod
      tags: [supplier-payment-methods]
      summary: Add a payment method to a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierPaymentMethodInput"
      responses:
        "201":
          description: Payment method created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPaymentMethod"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suppliers/{id}/payment-methods/{pmId}:
    put:
      operationId: updateSupplierPaymentMethod
      tags: [supplier-payment-methods]
      summary: Update a payment method
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pmId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierPaymentMethodInput"
      responses:
        "200":
          description: Updated payment method
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPaymentMethod"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSupplierPaymentMethod
      tags: [supplier-payment-methods]
      summary: Delete a payment method
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pmId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/SupplierCategory"
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - companyName
        - status
        - createdAt
        - categories
    CreateSupplierInput:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        categoryIds:
          type: array
          items:
            type: integer
      required:
        - companyName
    UpdateSupplierInput:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        status:
          type: string
        categoryIds:
          type: array
          items:
            type: integer
      required:
        - companyName
    SupplierPaymentMethod:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        type:
          type: string
        walletType:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - supplierId
        - type
        - createdAt
    CreateSupplierPaymentMethodInput:
      type: object
      properties:
        type:
          type: string
          description: "cash | wallet | instapay | bank"
        walletType:
          type: string
          description: "e.g. فودافون كاش, أورانج كاش — required when type=wallet"
        ownerName:
          type: string
        phone:
          type: string
          description: "Egyptian format +201XXXXXXXXX — required for wallet and instapay"
        bankName:
          type: string
          description: "required when type=bank"
        accountNumber:
          type: string
          description: "required when type=bank"
      required:
        - type
    SupplierCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success